requirements.txt
streamlit==1.45.0
pandas==2.2.3
plotly==6.0.1
langchain==0.3.25
langchain-openai==0.3.17
pydantic==2.11.3
python-dotenv==1.1.0
pytest==8.3.5
